creating a separate module for def helping tools
